Extract SDK executor secrets facade
